# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PandaWise API
  version: 0.2.0
  description: Sprint 1–2 contract for parent/child onboarding, discovery and GrowScore.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                required: [status, service, version]
                properties:
                  status: { const: ok }
                  service: { const: pandawise-api }
                  version: { type: string }
  /v1/config/bootstrap:
    get:
      operationId: getBootstrapConfiguration
      responses:
        "200":
          description: Public, non-sensitive master options needed by onboarding.
          content:
            application/json:
              schema:
                type: object
                required: [version, data]
                properties:
                  version: { type: string }
                  data:
                    $ref: "#/components/schemas/BootstrapData"
  /v1/auth/register:
    post:
      operationId: registerParent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterParentRequest"
      responses:
        "201":
          description: Parent account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400": { $ref: "#/components/responses/ValidationError" }
        "409": { $ref: "#/components/responses/ConflictError" }
  /v1/auth/login:
    post:
      operationId: loginParent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Parent authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
  /v1/auth/forgot-password:
    post:
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202":
          description: Generic acceptance response; does not disclose account existence.
  /v1/me:
    get:
      operationId: getCurrentParent
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current parent profile.
          content:
            application/json:
              schema:
                type: object
                required: [parent]
                properties:
                  parent: { $ref: "#/components/schemas/Parent" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
  /v1/children:
    get:
      operationId: listChildren
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Active children owned by the current parent.
          content:
            application/json:
              schema:
                type: object
                required: [children]
                properties:
                  children:
                    type: array
                    items: { $ref: "#/components/schemas/Child" }
    post:
      operationId: createChild
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChildRequest"
      responses:
        "201":
          description: Child profile created.
          content:
            application/json:
              schema:
                type: object
                required: [child]
                properties:
                  child: { $ref: "#/components/schemas/Child" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403":
          description: Plan child-profile limit reached.
  /v1/children/{childId}:
    get:
      operationId: getChild
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: childId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Child profile owned by the current parent.
          content:
            application/json:
              schema:
                type: object
                required: [child]
                properties:
                  child: { $ref: "#/components/schemas/Child" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /v1/children/{childId}/passions:
    parameters:
      - in: path
        name: childId
        required: true
        schema: { type: string }
    get:
      operationId: getChildPassions
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current non-scored Passion Discovery selections.
          content:
            application/json:
              schema:
                type: object
                required: [passionIds]
                properties:
                  passionIds:
                    type: array
                    items: { type: string }
    put:
      operationId: selectChildPassions
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [passionIds]
              properties:
                passionIds:
                  type: array
                  minItems: 1
                  maxItems: 5
                  uniqueItems: true
                  items: { type: string }
      responses:
        "200": { description: Passion selections saved. }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /v1/children/{childId}/assessments:
    post:
      operationId: startOrResumeAssessment
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: childId
          required: true
          schema: { type: string }
      responses:
        "201":
          description: A versioned Development Check was started or resumed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssessmentDetail" }
        "403": { description: Annual plan entitlement reached. }
        "409": { description: Passion Discovery is incomplete. }
  /v1/children/{childId}/growscore/latest:
    get:
      operationId: getLatestGrowScoreReport
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: childId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Latest completed GrowScore report for the child.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrowScoreReport" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /v1/assessments/{assessmentId}:
    get:
      operationId: getAssessment
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assessmentId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Development Check questions, selections and progress.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssessmentDetail" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /v1/assessments/{assessmentId}/responses/{questionId}:
    put:
      operationId: autoSaveAssessmentResponse
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assessmentId
          required: true
          schema: { type: string }
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [optionId]
              properties:
                optionId: { type: string }
                responseText: { type: string, maxLength: 500 }
      responses:
        "200": { description: Response validated, scored server-side and saved. }
        "409": { description: Assessment is already complete. }
  /v1/assessments/{assessmentId}/complete:
    post:
      operationId: completeAssessment
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assessmentId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: GrowScore calculated using scoring version 1.0.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrowScoreReport" }
        "409": { description: Required questions remain unanswered. }
  /v1/assessments/{assessmentId}/report:
    get:
      operationId: getGrowScoreReport
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: assessmentId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Plan-aware, strengths-first GrowScore report.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrowScoreReport" }
        "409": { description: Assessment is not complete. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnauthorizedError:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ConflictError:
      description: A conflicting account or resource already exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFoundError:
      description: The resource does not exist for this parent.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    Parent:
      type: object
      required:
        - id
        - name
        - parentType
        - mobileNumber
        - email
        - subscriptionPlanId
        - preferredLanguageId
        - dailyTimeCommitment
      properties:
        id: { type: string }
        name: { type: string }
        parentType:
          enum: [Mother, Father, Guardian, Grandparent]
        mobileNumber: { type: string }
        email: { type: string, format: email }
        subscriptionPlanId: { enum: [PLN001, PLN002, PLN003] }
        preferredLanguageId: { type: string }
        dailyTimeCommitment:
          $ref: "#/components/schemas/TimeCommitment"
    Child:
      type: object
      required:
        - id
        - parentId
        - name
        - dateOfBirth
        - ageYears
        - ageGroupId
        - gender
        - languageId
        - assessmentStatus
        - journeyStatus
        - currentBadgeLevel
        - currentStreak
      properties:
        id: { type: string }
        parentId: { type: string }
        name: { type: string }
        nickname: { type: string }
        avatarId: { type: string }
        dateOfBirth: { type: string, format: date }
        ageYears: { type: integer, minimum: 3, maximum: 12 }
        ageGroupId: { enum: [AG01, AG02, AG03] }
        gender: { enum: [Boy, Girl, Prefer Not to Say] }
        schoolId: { type: string }
        gradeId: { type: string }
        languageId: { type: string }
        assessmentStatus:
          enum: [Not Started, In Progress, Completed, Reassessment Due]
        journeyStatus: { enum: [Not Started, Active, Paused, Completed] }
        currentGrowScore: { type: number, minimum: 0, maximum: 100 }
        currentBadgeLevel: { enum: [Starter, Explorer, Achiever, Champion] }
        currentStreak: { type: integer, minimum: 0 }
    AuthResponse:
      type: object
      required: [token, parent]
      properties:
        token: { type: string }
        parent: { $ref: "#/components/schemas/Parent" }
    RegisterParentRequest:
      type: object
      required: [name, mobileNumber, email, password, termsAccepted]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        parentType:
          enum: [Mother, Father, Guardian, Grandparent]
          default: Guardian
        mobileNumber: { type: string, minLength: 8, maxLength: 20 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }
        preferredLanguageId: { type: string, default: LNG001 }
        dailyTimeCommitment:
          $ref: "#/components/schemas/TimeCommitment"
        termsAccepted: { const: true }
        marketingConsent: { type: boolean, default: false }
    CreateChildRequest:
      type: object
      required: [name, dateOfBirth, gender, languageId, parentTimeCommitment]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        nickname: { type: string, maxLength: 60 }
        avatarId: { type: string, maxLength: 80 }
        dateOfBirth: { type: string, format: date }
        gender: { enum: [Boy, Girl, Prefer Not to Say] }
        schoolId: { type: string }
        gradeId: { type: string }
        languageId: { type: string }
        knownInterests:
          type: array
          maxItems: 5
          items: { type: string }
        parentTimeCommitment:
          $ref: "#/components/schemas/TimeCommitment"
    BootstrapData:
      type: object
      required: [ageGroups, languages, schools, grades, skills, passions, timeCommitments]
      properties:
        ageGroups: { $ref: "#/components/schemas/MasterOptions" }
        languages: { $ref: "#/components/schemas/MasterOptions" }
        schools: { $ref: "#/components/schemas/MasterOptions" }
        grades: { $ref: "#/components/schemas/MasterOptions" }
        skills: { $ref: "#/components/schemas/MasterOptions" }
        passions: { $ref: "#/components/schemas/MasterOptions" }
        timeCommitments:
          type: array
          items: { $ref: "#/components/schemas/TimeCommitment" }
    MasterOptions:
      type: array
      items:
        type: object
        required: [id, name]
        properties:
          id: { type: string }
          name: { type: string }
        additionalProperties: true
    TimeCommitment:
      enum: [10_MIN, 15_MIN, 20_MIN, 30_MIN, WEEKENDS_ONLY]
    AssessmentDetail:
      type: object
      required: [assessment, child, questions, progress]
      properties:
        assessment:
          type: object
          required: [id, childId, version, depth, respondentMode, status, questionCount]
          properties:
            id: { type: string }
            childId: { type: string }
            version: { const: "1.0" }
            depth: { enum: [CORE, COMPREHENSIVE] }
            respondentMode: { enum: [PARENT, HYBRID] }
            status: { enum: [In Progress, Completed] }
            questionCount: { enum: [30, 50] }
        child:
          type: object
          additionalProperties: true
        questions:
          type: array
          items:
            type: object
            required: [id, skillId, text, respondentType, options]
            properties:
              id: { type: string }
              skillId: { type: string }
              text: { type: string }
              respondentType: { enum: [PARENT, CHILD] }
              selectedOptionId: { type: [string, "null"] }
              options:
                type: array
                items:
                  type: object
                  required: [id, text]
                  properties:
                    id: { type: string }
                    text: { type: string }
        progress:
          type: object
          required: [answered, total, percent]
          properties:
            answered: { type: integer, minimum: 0 }
            total: { enum: [30, 50] }
            percent: { type: integer, minimum: 0, maximum: 100 }
    GrowScoreSkill:
      type: object
      required: [skillId, name, score, band, bandLabel, message]
      properties:
        skillId: { type: string }
        name: { type: string }
        score: { type: number, minimum: 0, maximum: 100 }
        band:
          enum: [PRIORITY_GROWTH_AREA, DEVELOPING, AGE_APPROPRIATE, STRONG, EXCEPTIONAL]
        bandLabel: { type: string }
        message: { type: string }
    GrowScoreReport:
      type: object
      required:
        - assessment
        - child
        - growScore
        - scoreBand
        - scoreBandLabel
        - skills
        - strengths
        - recommendedFocusAreas
        - entitlements
      properties:
        assessment: { type: object, additionalProperties: true }
        child: { type: object, additionalProperties: true }
        growScore: { type: number, minimum: 0, maximum: 100 }
        scoreBand:
          enum: [PRIORITY_GROWTH_AREA, DEVELOPING, AGE_APPROPRIATE, STRONG, EXCEPTIONAL]
        scoreBandLabel: { type: string }
        skills:
          type: array
          items: { $ref: "#/components/schemas/GrowScoreSkill" }
        strengths:
          type: array
          maxItems: 3
          items: { $ref: "#/components/schemas/GrowScoreSkill" }
        recommendedFocusAreas:
          type: array
          maxItems: 3
          items: { $ref: "#/components/schemas/GrowScoreSkill" }
        entitlements:
          type: object
          required: [visibleSkillCount, lockedSkillCount, planId]
          properties:
            visibleSkillCount: { type: integer }
            lockedSkillCount: { type: integer }
            planId: { enum: [PLN001, PLN002, PLN003] }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array }